# Conflicts:
#	controllersvc/src/conf/controller-conf.xml
#	controllersvc/src/main/java/com/emc/storageos/computecontroller/impl/ComputeDeviceControllerImpl.java
#	controllersvc/src/main/java/com/emc/storageos/computecontroller/impl/ucs/UcsComputeDevice.java
#	controllersvc/src/main/java/com/emc/storageos/computesystemcontroller/impl/ComputeSystemControllerImpl.java
#	controllersvc/src/main/java/com/emc/storageos/filereplicationcontroller/FileReplicationDeviceController.java
#	controllersvc/src/main/java/com/emc/storageos/protectioncontroller/impl/recoverpoint/RPDeviceController.java
#	controllersvc/src/main/java/com/emc/storageos/vplexcontroller/VPlexDeviceController.java

+++ controllersvc/src/main/java/com/emc/storageos/computecontroller/impl/ucs/UcsDiscoveryWorker.java
@@ -21,15 +21,19 @@ import java.util.Set;
 
 import javax.xml.bind.JAXBElement;
 
+import com.google.common.base.Predicate;
+import com.google.common.collect.Collections2;
 import org.apache.commons.lang.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.emc.cloud.platform.ucs.out.model.BiosUnit;
 import com.emc.cloud.platform.ucs.out.model.ComputeBlade;
 import com.emc.cloud.platform.ucs.out.model.ComputeBoard;
 import com.emc.cloud.platform.ucs.out.model.FabricVlan;
 import com.emc.cloud.platform.ucs.out.model.FabricVsan;
 import com.emc.cloud.platform.ucs.out.model.FcPIo;
+import com.emc.cloud.platform.ucs.out.model.FirmwareRunning;
 import com.emc.cloud.platform.ucs.out.model.LsRequirement;
 import com.emc.cloud.platform.ucs.out.model.LsServer;
 import com.emc.cloud.platform.ucs.out.model.LsbootDef;
@@ -78,7 +82,9 @@ import com.emc.storageos.db.client.model.ComputeVnic;
 import com.emc.storageos.db.client.model.DataObject;
 import com.emc.storageos.db.client.model.DiscoveredDataObject;
 import com.emc.storageos.db.client.model.DiscoveredSystemObject;
+import com.emc.storageos.db.client.model.Host;
 import com.emc.storageos.db.client.model.StringSet;
+import com.emc.storageos.db.client.model.UCSServiceProfile;
 import com.emc.storageos.db.client.model.UCSServiceProfileTemplate;
 import com.emc.storageos.db.client.model.UCSVhbaTemplate;
 import com.emc.storageos.db.client.model.UCSVnicTemplate;
@@ -98,6 +104,7 @@ public class UcsDiscoveryWorker {
     private static final String BLADE_REMOVED = "removed";
     private static final String BLADE_CFG_FAILURE = "config-failure";
     private static final String BLADE_AVAILABLE = "available";
+    private static final String DERIVED = "derived";
 
     private UCSMService ucsmService;
     private DbClient _dbClient;
@@ -143,6 +150,7 @@ public class UcsDiscoveryWorker {
         URL ucsmURL = getUcsmURL(cs);
 
         List<ComputeBlade> computeBlades;
+        List<LsServer> allServiceProfiles;
         Map<String, LsServer> associatedLsServers;
         List<LsServer> serviceProfileTemplates;
         List<VnicLanConnTempl> vnicTemplates;
@@ -160,6 +168,7 @@ public class UcsDiscoveryWorker {
             verifyVersion(cs, ucsmVersion);
 
             computeBlades = ucsmService.getComputeBlades(ucsmURL.toString(), cs.getUsername(), cs.getPassword());
+            allServiceProfiles = ucsmService.getAllServiceProfiles(ucsmURL.toString(), cs.getUsername(), cs.getPassword());
             associatedLsServers = ucsmService.getAllAssociatedLsServers(ucsmURL.toString(), cs.getUsername(), cs.getPassword());
             serviceProfileTemplates = ucsmService.getServiceProfileTemplates(ucsmURL.toString(), cs.getUsername(), cs.getPassword());
             bootPolicies = ucsmService.getBootPolicies(ucsmURL.toString(), cs.getUsername(), cs.getPassword());
@@ -179,32 +188,38 @@ public class UcsDiscoveryWorker {
             } else {
                 cs.setLastDiscoveryStatusMessage(e.getMessage());
             }
-            _dbClient.persistObject(cs);
+            _dbClient.updateObject(cs);
             throw ComputeSystemControllerException.exceptions.discoverFailed(computeSystemURI.toString(), e);
         }
+        try{
+            reconcileServiceProfileTemplates(cs, serviceProfileTemplates);
+            reconcileServiceProfiles(cs,allServiceProfiles);
+            reconcileComputeBlades(cs, computeBlades, associatedLsServers);
+            reconcileVhbas(cs, associatedLsServers, new VhbaHelper(vsanFabricList));
+            reconcileServiceProfileTemplatesHBAs(cs, serviceProfileTemplates, new VhbaHelper(vsanFabricList));
+            reconcileServiceProfileTemplatesVnics(cs, serviceProfileTemplates);
+            reconcileServiceProfileTemplatesBootDefinitions(cs, serviceProfileTemplates);
+            reconcileBootPolicies(cs, bootPolicies);
+            reconcileVnicTemplates(cs, vnicTemplates);
+            reconcileVhbaTemplates(cs, vhbaTemplates);
+    
+            Map<String, Set<String>> unpinnedVsans = getUnpinnedVSans(vsanList, fcInterfaceMap);
+            reconcileUplinkPorts(cs, uplinkMap, fcInterfaceMap, unpinnedVsans);
+            reconcileUplinkPortChannels(cs, portChannelMap, unpinnedVsans);
+            reconcileVlans(cs, vlanList);
+    
+            associateComputeImageServer(cs);
+            matchComputeBladesToHosts(cs);
 
-        reconcileServiceProfileTemplates(cs, serviceProfileTemplates);
-        reconcileComputeBlades(cs, computeBlades, associatedLsServers);
-        reconcileVhbas(cs, associatedLsServers, new VhbaHelper(vsanFabricList));
-        reconcileServiceProfileTemplatesHBAs(cs, serviceProfileTemplates, new VhbaHelper(vsanFabricList));
-        reconcileServiceProfileTemplatesVnics(cs, serviceProfileTemplates);
-        reconcileServiceProfileTemplatesBootDefinitions(cs, serviceProfileTemplates);
-        reconcileBootPolicies(cs, bootPolicies);
-        reconcileVnicTemplates(cs, vnicTemplates);
-        reconcileVhbaTemplates(cs, vhbaTemplates);
-
-        Map<String, Set<String>> unpinnedVsans = getUnpinnedVSans(vsanList, fcInterfaceMap);
-        reconcileUplinkPorts(cs, uplinkMap, fcInterfaceMap, unpinnedVsans);
-        reconcileUplinkPortChannels(cs, portChannelMap, unpinnedVsans);
-        reconcileVlans(cs, vlanList);
-
-        matchComputeBladesToHosts(cs);
-
-        cs.setLastDiscoveryRunTime(Calendar.getInstance().getTimeInMillis());
-        cs.setSuccessDiscoveryTime(Calendar.getInstance().getTimeInMillis());
-        cs.setDiscoveryStatus(DiscoveredDataObject.DataCollectionJobStatus.COMPLETE.name());
-        associateComputeImageServer(cs);
-        _dbClient.updateObject(cs);
+            cs.setLastDiscoveryRunTime(Calendar.getInstance().getTimeInMillis());
+            cs.setSuccessDiscoveryTime(Calendar.getInstance().getTimeInMillis());
+            cs.setDiscoveryStatus(DiscoveredDataObject.DataCollectionJobStatus.COMPLETE.name());
+        } catch (ComputeSystemControllerException e){
+            cs.setLastDiscoveryStatusMessage(e.getMessage());
+            throw ComputeSystemControllerException.exceptions.discoverFailed(cs.getId().toString(), e);
+        } finally {
+           _dbClient.persistObject(cs);
+        }
     }
 
     private void verifyVersion(ComputeSystem cs, String version) {
@@ -337,10 +352,14 @@ public class UcsDiscoveryWorker {
         createDataObjects(new ArrayList<DataObject>(addBlades.values()));
         persistDataObjects(new ArrayList<DataObject>(updateBlades.values()));
 
-        for (String name : removeBlades.keySet()) {
-            _log.info("Marked for deletion ComputeElement name:" + name);
+        if (!removeBlades.isEmpty()){
+            for (String name : removeBlades.keySet()) {
+                _log.info("Marked for deletion ComputeElement name:" + name);
+            }
+            removeBladesFromComputeVirtualPools(removeBlades.values());
+            removeBladesFromHosts(removeBlades.values());
+            deleteDataObjects(new ArrayList<DataObject>(removeBlades.values()));
         }
-        deleteDataObjects(new ArrayList<DataObject>(removeBlades.values()));
     }
 
     private void createComputeElement(ComputeSystem cs, ComputeElement computeElement, ComputeBlade computeBlade, LsServer lsServer) {
@@ -381,6 +400,7 @@ public class UcsDiscoveryWorker {
         computeElement.setLastDiscoveryRunTime(Calendar.getInstance().getTimeInMillis());
         computeElement.setSuccessDiscoveryTime(Calendar.getInstance().getTimeInMillis());
         computeElement.setDiscoveryStatus(DiscoveredDataObject.DataCollectionJobStatus.COMPLETE.name());
+        computeElement.setBios(getBios(computeBlade));
 
         if (lsServer != null) {
             computeElement.setAvailable(false);
@@ -398,7 +418,127 @@ public class UcsDiscoveryWorker {
             computeElement.setUuid(computeBlade.getUuid());
             computeElement.setDn(NullColumnValueGetter.getNullStr());
         }
+
     }
+
+    private static String getBios(ComputeBlade computeBlade) {
+        for(Serializable bladeContent : computeBlade.getContent()){
+            if((bladeContent != null) && (bladeContent instanceof JAXBElement<?>) &&
+                    (((JAXBElement)bladeContent).getValue() instanceof BiosUnit)) {
+                BiosUnit biosUnit = (BiosUnit) ((JAXBElement)bladeContent).getValue();
+                for(Serializable biosUnitContent : biosUnit.getContent()) {
+                    if((biosUnitContent != null) && (biosUnitContent instanceof JAXBElement<?>) &&
+                            (((JAXBElement)biosUnitContent).getValue() instanceof FirmwareRunning)) {
+                        FirmwareRunning firmwareRunning = (FirmwareRunning) ((JAXBElement)biosUnitContent).getValue();
+                        return firmwareRunning.getVersion();
+                    }
+                }
+            }
+        }
+        return null;
+    }
+
+    private void reconcileServiceProfiles(ComputeSystem cs, List<LsServer> allLsServers){
+        _log.info("Reconciling UCS Service Profiles");
+
+        URIQueryResultList uris = new URIQueryResultList();
+        _dbClient.queryByConstraint(ContainmentConstraint.Factory
+                .getComputeSystemServiceProfilesConstraint(cs.getId()), uris);
+
+        Map<String, UCSServiceProfile> removeServiceProfiles = new HashMap<>();
+        Map<String, UCSServiceProfile> updateServiceProfiles = new HashMap<>();
+        Map<String, UCSServiceProfile> addServiceProfiles = new HashMap<>();
+
+        List<UCSServiceProfile> serviceProfiles = _dbClient.queryObject(UCSServiceProfile.class, uris, true);
+
+        for (UCSServiceProfile serviceProfile : serviceProfiles) {
+            removeServiceProfiles.put(serviceProfile.getDn(), serviceProfile);
+        }
+
+        // discovered data
+        for (LsServer lsServer : filterLsServers(allLsServers)) {
+            UCSServiceProfile serviceProfile = removeServiceProfiles.get(lsServer.getDn());
+            if (serviceProfile != null) {
+                removeServiceProfiles.remove(lsServer.getDn());
+                updateUCSServiceProfile(serviceProfile, lsServer);
+                updateServiceProfiles.put(lsServer.getDn(), serviceProfile);
+            } else {
+                serviceProfile = new UCSServiceProfile();
+                createUCSServiceProfile(cs, serviceProfile, lsServer);
+                addServiceProfiles.put(lsServer.getDn(), serviceProfile);
+            }
+        }
+        createDataObjects(new ArrayList<DataObject>(addServiceProfiles.values()));
+        persistDataObjects(new ArrayList<DataObject>(updateServiceProfiles.values()));
+
+        if (!removeServiceProfiles.isEmpty()){
+            for (String key : removeServiceProfiles.keySet()) {
+               _log.info("Marked for deletion UCSServiceProfile: " + key);
+            }   
+            
+            removeServiceProfilesFromHosts(removeServiceProfiles.values());
+            deleteDataObjects(new ArrayList<DataObject>(removeServiceProfiles.values()));
+        }
+        validateServiceProfileUuids(cs);
+
+    }
+
+    private void validateServiceProfileUuids(ComputeSystem cs){
+        URIQueryResultList uris = new URIQueryResultList();
+        _dbClient.queryByConstraint(ContainmentConstraint.Factory
+                .getComputeSystemServiceProfilesConstraint(cs.getId()), uris);
+
+        List<UCSServiceProfile> serviceProfiles = _dbClient.queryObject(UCSServiceProfile.class, uris, true);
+        Map<String,UCSServiceProfile> uuidMap = new HashMap<>();
+        for (UCSServiceProfile serviceProfile : serviceProfiles) {
+            UCSServiceProfile anotherProfile = uuidMap.get(serviceProfile.getUuid());
+            if (anotherProfile == null) {
+                uuidMap.put(serviceProfile.getUuid(), serviceProfile);
+            }else {
+                _log.info("Found two service profiles {} , {}  that have same uuid: {} ", serviceProfile.getDn(), anotherProfile.getDn(), serviceProfile.getUuid());
+                throw ComputeSystemControllerException.exceptions.serviceProfileUuidDuplicate(serviceProfile.getDn(), anotherProfile.getDn(), serviceProfile.getUuid());
+            }
+        }
+    }
+
+    private void createUCSServiceProfile(ComputeSystem cs, UCSServiceProfile serviceProfile, LsServer lsServer) {
+
+        _log.info("Adding UCSServiceProfile label: " + lsServer.getDn());
+        URI uri = URIUtil.createId(UCSServiceProfile.class);
+        serviceProfile.setComputeSystem(cs.getId());
+        serviceProfile.setInactive(false);
+        serviceProfile.setId(uri);
+        serviceProfile.setSystemType(cs.getSystemType());
+        serviceProfile.setCreationTime(Calendar.getInstance());
+        serviceProfile.setDn(lsServer.getDn());
+        serviceProfile.setNativeGuid(NativeGUIDGenerator.generateNativeGuid(cs, serviceProfile));
+        updateUCSServiceProfile(serviceProfile, lsServer);
+    }
+
+    private void updateUCSServiceProfile(UCSServiceProfile serviceProfile, LsServer lsServer) {
+
+        _log.info("Updating UCSServiceProfile id: " + serviceProfile.getId());
+        serviceProfile.setLabel(lsServer.getName());
+        serviceProfile.setComputeElementDn(lsServer.getPnDn());
+        // Fail discovery if the uuid of a service profile matching a managed host changes!
+        if (serviceProfile.getUuid() == null) {
+            serviceProfile.setUuid(lsServer.getUuid());
+        }else if (!serviceProfile.getUuid().equals(lsServer.getUuid()) ){
+            if ( !NullColumnValueGetter.isNullURI(serviceProfile.getHost())){
+                String errorMessage = "uuid of service profile" + lsServer.getDn() + " changed from : "+ serviceProfile.getUuid() + " to : "+ lsServer.getUuid();
+                _log.error(errorMessage);
+                throw ComputeSystemControllerException.exceptions.serviceProfileUuidChanged(lsServer.getDn(), serviceProfile.getUuid(), lsServer.getUuid());
+            }else {
+                serviceProfile.setUuid(lsServer.getUuid());
+            }
+        }
+
+        serviceProfile.setLastDiscoveryRunTime(Calendar.getInstance().getTimeInMillis());
+        serviceProfile.setSuccessDiscoveryTime(Calendar.getInstance().getTimeInMillis());
+        serviceProfile.setDiscoveryStatus(DiscoveredDataObject.DataCollectionJobStatus.COMPLETE.name());
+        serviceProfile.setCompatibilityStatus(DiscoveredDataObject.CompatibilityStatus.UNKNOWN.name());
+    }
+
 
     private void reconcileVhbas(ComputeSystem cs, Map<String, LsServer> associatedLsServers, VhbaHelper lookUpVsan) {
         _log.info("Reconciling Vhbas");
@@ -1644,6 +1784,68 @@ public class UcsDiscoveryWorker {
         computeVnic.setNativeVlan(nativeVlan);
         computeVnic.setVlans(vlans);
     }
+    private void removeServiceProfilesFromHosts(Collection<UCSServiceProfile> serviceProfiles) {
+         List<UCSServiceProfile> serviceProfilesToUpdate = new ArrayList<UCSServiceProfile>();
+         for (UCSServiceProfile serviceProfile : serviceProfiles) {
+             if (!NullColumnValueGetter.isNullURI(serviceProfile.getHost())){
+                Host host = _dbClient.queryObject(Host.class,serviceProfile.getHost());
+                if (host!=null){
+                     _log.info("Removing UCSServiceProfile {} association from Host {} ", serviceProfile.getDn(), host.getLabel());
+                     host.setServiceProfile(NullColumnValueGetter.getNullURI());
+                    _dbClient.persistObject(host);
+                }
+                _log.info("Removing Host association from service profile {}",  serviceProfile.getDn());
+                serviceProfile.setHost(NullColumnValueGetter.getNullURI());
+                serviceProfilesToUpdate.add(serviceProfile);
+             }
+        }
+        if (!serviceProfilesToUpdate.isEmpty()){
+            persistDataObjects(new ArrayList<DataObject>(serviceProfilesToUpdate));
+        }
+
+    }
+    private void removeBladesFromHosts(Collection<ComputeElement> removeBlades) {
+        List<URI> ids = _dbClient.queryByType(Host.class, true);
+        Iterator<Host> iter = _dbClient.queryIterativeObjects(Host.class, ids);
+
+        while (iter.hasNext()) {
+            Host host = iter.next();
+            for (ComputeElement computeElement : removeBlades) {
+                if (host.getComputeElement() != null
+                        && host.getComputeElement().equals(computeElement.getId())) {
+                    _log.info("Removing ComputeElement {} association from Host {} ", computeElement.getDn(), host.getLabel());
+                    host.setComputeElement(NullColumnValueGetter.getNullURI());
+                    _dbClient.persistObject(host);
+                    break;
+                }
+            }
+
+        }
+    }
+
+
+    private void removeBladesFromComputeVirtualPools(Collection<ComputeElement> removeBlades) {
+        List<URI> ids = _dbClient.queryByType(ComputeVirtualPool.class, true);
+        Iterator<ComputeVirtualPool> iter = _dbClient.queryIterativeObjects(ComputeVirtualPool.class, ids);
+
+        while (iter.hasNext()) {
+            Boolean dbUpdateRequired = false;
+            ComputeVirtualPool cvp = iter.next();
+            for (ComputeElement computeElement : removeBlades) {
+                if (cvp.getMatchedComputeElements() != null
+                        && cvp.getMatchedComputeElements().contains(computeElement.getId().toString())) {
+                    _log.info("Removing ComputeElement {} from ComputeVirtualPool {} ", computeElement.getDn(), cvp.getLabel());
+                    cvp.removeMatchedComputeElement(computeElement.getId().toString());
+                    dbUpdateRequired = true;
+                }
+            }
+
+            if (dbUpdateRequired) {
+                _log.info("Persisting ComputeVirtualPool {},after ComputeElement removal", cvp.getLabel());
+                _dbClient.persistObject(cvp);
+            }
+        }
+    }
 
     private void removeServiceProfileTemplatesFromComputeVirtualPool(Collection<UCSServiceProfileTemplate> removeTemplates) {
         List<URI> ids = _dbClient.queryByType(ComputeVirtualPool.class, true);
@@ -2060,11 +2262,8 @@ public class UcsDiscoveryWorker {
         return new Integer(0);
     }
 
-    private void matchComputeBladesToHosts(ComputeSystem cs) {
-        URIQueryResultList uris = new URIQueryResultList();
-        _dbClient.queryByConstraint(ContainmentConstraint.Factory
-                .getComputeSystemComputeElemetsConstraint(cs.getId()), uris);
-        HostToComputeElementMatcher.matchComputeElementsToHostsByUuid(uris, _dbClient);
+    private void matchComputeBladesToHosts(ComputeSystem computeSystem) {
+        HostToComputeElementMatcher.matchAllHostsToComputeElements(_dbClient, computeSystem.getId());
     }
 
     /**
@@ -2099,6 +2298,22 @@ public class UcsDiscoveryWorker {
                 }
             }
         }
+    }
+
+    /**
+     * Filter the list of LsServers.
+     *
+     * @param allLsServers  List of LsServer instances.
+     * @return              Filtered List view of allLsServers.
+     */
+    private Collection<LsServer> filterLsServers(List<LsServer> allLsServers) {
+        return Collections2.filter(allLsServers, new Predicate<LsServer>() {
+            @Override
+            public boolean apply(LsServer lsServer) {
+                // Filter out LsServer instances with derived / hardware-default UUIDs.
+                return !lsServer.getUuid().equalsIgnoreCase(DERIVED);
+            }
+        });
     }
 
 }
